Json writer rewritten to use encoder.  And test of scripts completed

+++ output/normalize.go
@@ -8,7 +8,7 @@ func NormalizeScriptMFCC(structs []Script, numMFCC int) []Script {
 		var count float64
 		for _, str := range structs {
 			for _, mf := range str.MFCC {
-				sum += float64(mf[col])
+				sum += mf[col]
 				count++
 			}
 		}
@@ -16,15 +16,15 @@ func NormalizeScriptMFCC(structs []Script, numMFCC int) []Script {
 		var devSqr float64
 		for _, str := range structs {
 			for _, mf := range str.MFCC {
-				devSqr += math.Pow(float64(mf[col])-mean, 2)
+				devSqr += math.Pow(mf[col]-mean, 2)
 			}
 		}
 		var stddev = math.Sqrt(devSqr / count)
 		for i, str := range structs {
 			var mfccs = str.MFCC
 			for j, mf := range mfccs {
-				value := float64(mf[col])
-				mfccs[j][col] = float32((value - mean) / stddev)
+				value := mf[col]
+				mfccs[j][col] = (value - mean) / stddev
 			}
 			structs[i].MFCC = mfccs
 		}
@@ -69,7 +69,7 @@ func PadScriptRows(structs []Script, numMFCC int) []Script {
 			largest = str.MFCCRows
 		}
 	}
-	var padRow = make([]float32, numMFCC)
+	var padRow = make([]float64, numMFCC)
 	for i, str := range structs {
 		mfccs := str.MFCC
 		needRows := largest - str.MFCCRows
